moved discount interaction to discount controller
